requirements.txt
google-auth==2.6.6
google-auth-oauthlib==0.4.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
nltk==3.7
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.7.0
rake-nltk==1.0.6
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sklearn==0.0
streamlit==1.8.1